# Speed up frame masking/unmasking with C extension