--- a/tox.ini
+++ b/tox.ini
@@ -1,17 +1,28 @@
 [tox]
-envlist = py33,py34,py35,py36,coverage,flake8,isort
+envlist = {py33,py34,py35,py36}{,-cext},coverage-cext-{win32,linux,darwin},flake8,isort
 
 [testenv]
+platform =
+    win32: win32
+    linux: linux
+    darwin: darwin
 deps =
     py33: asyncio
-commands = python -m unittest
-
-[testenv:coverage]
+    coverage: coverage
+    coverage-{linux,darwin}: gcovr
+setenv =
+    coverage-{linux,darwin}: CFLAGS=-O0 -coverage
+    coverage-{linux,darwin}: LDFLAGS=-coverage
 commands =
-    python -m coverage erase
-    python -m coverage run --branch --source=websockets -m unittest
-    python -m coverage report --show-missing --fail-under=100
-deps = coverage
+               python setup.py cleanext clean --all
+    cext:      python setup.py build_ext --inplace
+               python -m unittest
+    coverage:  python -m coverage erase
+    coverage:  python -m coverage run --branch --source=websockets -m unittest
+    coverage:  python setup.py cleanext
+    coverage:  python -m coverage run -a --branch --source=websockets -m unittest
+    coverage:  python -m coverage report --show-missing --fail-under=100
+    coverage-{linux,darwin}: gcovr -r . -s #waiting for fail under option
 
 [testenv:flake8]
 commands = flake8 websockets