A small benchmark gives these figures:
```
pure_python_tobytearray:  12774.686 MB/s (6,291,456 bytes)
pure_python_old:              5.193 MB/s (6,291,456 bytes)
pure_python_new:              7.701 MB/s (6,291,456 bytes)
apply_mask:               10773.711 MB/s (6,291,456 bytes)
pure_python_tobytearray:  21022.988 MB/s (60,000 bytes)
pure_python_old:              5.135 MB/s (60,000 bytes)
pure_python_new:              7.711 MB/s (60,000 bytes)
apply_mask:               22086.288 MB/s (60,000 bytes)
pure_python_tobytearray:     16.562 MB/s (6 bytes)
pure_python_old:              2.831 MB/s (6 bytes)
pure_python_new:              2.954 MB/s (6 bytes)
apply_mask:                  23.026 MB/s (6 bytes)
pure_python_tobytearray:      2.734 MB/s (1 bytes)
pure_python_old:              0.867 MB/s (1 bytes)
pure_python_new:              0.835 MB/s (1 bytes)
apply_mask:                   4.002 MB/s (1 bytes)
```
Core i7-4870HQ @ 2.50GHz,
Apple LLVM version 8.1.0 (clang-802.0.42),
Python 3.6.0

Similar to aaugustin/websockets#175 PR