@mayeut 2017-07-21
Here are the results for the Autobahn Testsuite (wstest running Python 2.7.10):
```
WEBSOCKETS_CEXT_DISABLED=1 python test_server.py
time wstest -m fuzzingclient
real 0m45.950s
user 0m10.202s
sys  0m1.667s

python test_server.py
time wstest -m fuzzingclient
real 0m25.551s
user 0m10.272s
sys  0m1.650s

wstest -m fuzzingserver
WEBSOCKETS_CEXT_DISABLED=1 time python test_client.py
real 0m44.640s
user 0m27.666s
sys  0m1.513s

wstest -m fuzzingserver
time python test_client.py
real 0m23.939s
user 0m7.386s
sys 0m1.128s
```