@mayeut 2017-07-21
Here is a benchmark from a client/server I'm working on.
The client sends a message that's around 256kB to the server which responds with a few kB (between 1kB & 64kB).
The server is multithreaded and processes long operations (3 kinds here) in a C extension outside Python GIL (Global Interpreter Lock).

Before PR
```
slowest operation => 10.3 operations/second
normal operation  => 15.4 operations/second
fastest operation => 17.9 operations/second
```

After PR
```
slowest operation =>  14.0 operations/second
normal operation  =>  62.5 operations/second
fastest operation => 933.9 operations/second
```