@aaugustin 2017-07-22
I'm working on this. Answering my questions above would be very useful. It's probably best not to write more code until we agree on what we want.