@mayeut 2017-07-22
PS:
Here's the code I use for benchmarking:

```
#!/usr/bin/python
# coding: utf-8

from itertools import cycle
import timeit

def pure_python_old(mask_bits, data):
    return bytes(b ^ mask_bits[i % 4] for i, b in enumerate(data))

def pure_python_new(mask_bits, data):
    return bytes(b ^ m for b, m in zip(data, cycle(mask_bits)))

def pure_python_tobytearray(mask_bits, data):
    return bytearray(data)

try:
    from websockets.cext import apply_mask
except ImportError as err:
    print(err)
    def apply_mask(mask, data):
        return pure_python_new(mask, data)

mask_bits = b'L\xde\xcb:'

def fn4(fn, data):
    data = fn(mask_bits, data)
    data = fn(mask_bits, data)
    data = fn(mask_bits, data)
    return fn(mask_bits, data)

def mytimeit(fn, data):
    number = 0
    start_time = timeit.default_timer()
    while True:
        data = fn4(fn, data)
        number += 4
        stop_time = timeit.default_timer()
        if stop_time - start_time > 1.0:
            break
    iter = number
    start_time = timeit.default_timer()
    while iter > 0:
        fn4(fn, data)
        iter -= 4
    stop_time = timeit.default_timer()
    print('{0:<25s} {1:9.3f} MB/s ({2:,d} bytes)'.format(fn.__name__ + ':', ((number * len(data)) / (1024.0 * 1024.0)) / (stop_time - start_time), len(data)))

def bench(data):
    mytimeit(pure_python_tobytearray, data)
    mytimeit(pure_python_old, data)
    mytimeit(pure_python_new, data)
    mytimeit(apply_mask, data)

bench(b'abcdef' * (1024 * 1024))
bench(b'abcdef' * 10000)
bench(b'abcdef')
bench(b'a')
```