@cjerdonek 2017-07-22
> We can't ship a universal wheel anymore. Is there a service that can build wheels on various common OSes for us?

Do you already know about this: https://github.com/pypa/manylinux ? While it doesn't cover every common OS, it may cover the 80% case.