@mayeut 2017-07-22
> We can't ship a universal wheel anymore. Is there a service that can build wheels on various common OSes for us?

As @cjerdonek said, wheels for linux can be built by any service that allows the use of docker by using the `manylinux1_x86_64` and `manylinux1_i686`
Wheels for Windows can be built using the [appveyor](https://www.appveyor.com) service, at least for CPython 3.5 and 3.6 (given available compilers and [CPython requirements](https://wiki.python.org/moin/WindowsCompilers#Which_Microsoft_Visual_C.2B-.2B-_compiler_to_use_with_a_specific_Python_version_.3F)
MacOs wheels can be built using travis-ci
c.f. also https://github.com/MacPython/wiki/wiki/Wheel-building