@aaugustin 2017-07-22
Thanks for your comments. I'll review them tomorrow and hopefully merge #214.

Then I'll take a look at optimizing performance with SSE. I hoped the compiler would be able to figure that out by itself but disassembling the .so shows that it doesn't...