@aaugustin 2017-07-23
I integrated all the parts I'm comfortable with in #214.

The other parts are:

1. support for cleaning up the compiled extension in setup.py: I really hate having a non-trivial setup.py, I regretted that too many times in the past, so I'll only do this if it solves a big pain point.

2. coverage for the C code: this is something I want, but it depends on 1. which I don't have the energy to investigate right now

3. support for running tox on Windows: mostly useful if we try to run CI on Windows with Appveyor

4. when not doing the XOR in 16 bytes blocks with SSE2, doing it in 8 bytes blocks: this looks fairly easy to implement and could be useful on non x86 CPUs (I assume that every x86 CPU worth caring about has SSE2 at this point)
    - there's also a 4 bytes block done once, I don't think that's worth the special case, the overhead of the extra conditional is likely higher than looping 4 more times in the byte-by-byte loop.